+++ src/simsopt/geo/__init__.py
@@ -22,6 +22,7 @@ from .surfacehenneberg import *
 from .surfaceobjectives import *
 from .surfacerzfourier import *
 from .surfacexyzfourier import *
+from .surfacespline import *
 from .surfacexyztensorfourier import *
 from .strain_optimization import *
 from .wireframe_toroidal import *
@@ -39,6 +40,7 @@ __all__ = (curve.__all__ + curvehelical.__all__ +
            surface.__all__ +
            surfacegarabedian.__all__ + surfacehenneberg.__all__ +
            surfacerzfourier.__all__ + surfacexyzfourier.__all__ +
+           surfacespline.__all__ + 
            surfacexyztensorfourier.__all__ + surfaceobjectives.__all__ +
            strain_optimization.__all__ + framedcurve.__all__ +
            wireframe_toroidal.__all__ + ports.__all__ +
